Refactor expo-image for web
