+++ packages/expo-image/src/web/ImageWrapper.tsx
@@ -1,54 +1,14 @@
-import React, { CSSProperties, SyntheticEvent, useEffect, Ref, useMemo } from 'react';
+import React, { useEffect, Ref, useMemo } from 'react';
 
-import {
-  ImageContentPositionObject,
-  ImageContentPositionValue,
-  ImageNativeProps,
-  ImageSource,
-} from '../Image.types';
+import { ImageNativeProps, ImageSource } from '../Image.types';
 import { useBlurhash } from '../utils/blurhash/useBlurhash';
 import { isBlurhashString, isThumbhashString } from '../utils/resolveSources';
 import { thumbHashStringToDataURL } from '../utils/thumbhash/thumbhash';
+import ColorTintFilter, { getColorTintStyle } from './ColorTintFilter';
+import { ImageWrapperProps } from './ImageWrapper.types';
+import { getImageWrapperEventHandler } from './getImageWrapperEventHandler';
+import { absoluteFilledPosition, getObjectPositionFromContentPositionObject } from './positioning';
 import { SrcSetSource } from './useSourceSelection';
-
-function ensureUnit(value: string | number) {
-  const trimmedValue = String(value).trim();
-  if (trimmedValue.endsWith('%')) {
-    return trimmedValue;
-  }
-  return `${trimmedValue}px`;
-}
-
-type KeysOfUnion<T> = T extends T ? keyof T : never;
-
-function getObjectPositionFromContentPositionObject(
-  contentPosition?: ImageContentPositionObject
-): string {
-  const resolvedPosition = { ...contentPosition } as Record<
-    KeysOfUnion<ImageContentPositionObject>,
-    ImageContentPositionValue
-  >;
-  if (!resolvedPosition) {
-    return '50% 50%';
-  }
-  if (resolvedPosition.top == null && resolvedPosition.bottom == null) {
-    resolvedPosition.top = '50%';
-  }
-  if (resolvedPosition.left == null && resolvedPosition.right == null) {
-    resolvedPosition.left = '50%';
-  }
-
-  return (
-    ['top', 'bottom', 'left', 'right']
-      .map((key) => {
-        if (key in resolvedPosition) {
-          return `${key} ${ensureUnit(resolvedPosition[key])}`;
-        }
-        return '';
-      })
-      .join(' ') || '50% 50%'
-  );
-}
 
 function getFetchPriorityFromImagePriority(priority: ImageNativeProps['priority'] = 'normal') {
   return priority && ['low', 'high'].includes(priority) ? priority : 'auto';
@@ -78,28 +38,13 @@ const ImageWrapper = React.forwardRef(
       className,
       accessibilityLabel,
       ...props
-    }: {
-      source?: ImageSource | SrcSetSource | null;
-      events?: {
-        onLoad?: (((event: SyntheticEvent<HTMLImageElement, Event>) => void) | undefined | null)[];
-        onError?: ((({ source }: { source: ImageSource | null }) => void) | undefined | null)[];
-        onTransitionEnd?: ((() => void) | undefined | null)[];
-        onMount?: ((() => void) | undefined | null)[];
-      };
-      contentPosition?: ImageContentPositionObject;
-      hashPlaceholderContentPosition?: ImageContentPositionObject;
-      priority?: string | null;
-      style: CSSProperties;
-      tintColor?: string | null;
-      hashPlaceholderStyle?: CSSProperties;
-      className?: string;
-      accessibilityLabel?: string;
-    },
+    }: ImageWrapperProps,
     ref: Ref<HTMLImageElement>
   ) => {
     useEffect(() => {
       events?.onMount?.forEach((e) => e?.());
     }, []);
+
     const isBlurhash = isBlurhashString(source?.uri || '');
     const isThumbhash = isThumbhashString(source?.uri || '');
     const isHash = isBlurhash || isThumbhash;
@@ -121,52 +66,29 @@ const ImageWrapper = React.forwardRef(
     );
 
     const uri = isHash ? blurhashUri ?? thumbhashUri : source?.uri;
+
     return (
       <>
-        {tintColor && (
-          <svg>
-            <defs>
-              <filter id={`tint-${tintColor}`} x="0" y="0" width="0" height="0">
-                <feFlood floodColor={tintColor} floodOpacity="1" result="flood" />
-                <feComposite in="flood" in2="SourceAlpha" operator="in" />
-              </filter>
-            </defs>
-          </svg>
-        )}
+        <ColorTintFilter tintColor={tintColor} />
         <img
           ref={ref}
           alt={accessibilityLabel}
           className={className}
           src={uri || undefined}
-          {...getImgPropsFromSource(source)}
           key={source?.uri}
-          {...props}
           style={{
-            width: '100%',
-            height: '100%',
-            position: 'absolute',
-            left: 0,
-            right: 0,
             objectPosition,
-            filter: tintColor ? `url(#tint-${tintColor})` : '',
-            ...style,
+            ...absoluteFilledPosition,
+            ...getColorTintStyle(tintColor),
             ...(isHash ? hashPlaceholderStyle : {}),
+            ...style,
           }}
           // @ts-ignore
           // eslint-disable-next-line react/no-unknown-property
           fetchpriority={getFetchPriorityFromImagePriority(priority || 'normal')}
-          onLoad={(event) => {
-            if (typeof window !== 'undefined') {
-              // this ensures the animation will run, since the starting class is applied at least 1 frame before the target class set in the onLoad event callback
-              window.requestAnimationFrame(() => {
-                events?.onLoad?.forEach((e) => e?.(event));
-              });
-            } else {
-              events?.onLoad?.forEach((e) => e?.(event));
-            }
-          }}
-          onTransitionEnd={() => events?.onTransitionEnd?.forEach((e) => e?.())}
-          onError={() => events?.onError?.forEach((e) => e?.({ source: source || null }))}
+          {...getImageWrapperEventHandler(events, source)}
+          {...getImgPropsFromSource(source)}
+          {...props}
         />
       </>
     );
