+++ packages/expo-image/src/web/imageStyles.tsx
@@ -2,6 +2,8 @@ const style = `
 [data-expoimage] .cross-dissolve {
   transition-property: opacity;
   animation-fill-mode: forwards;
+  -webkit-transform: translate3d(0,0,0);
+    transform: translate3d(0,0,0);
 }
 [data-expoimage] .cross-dissolve-start:not(.transitioning) {
   opacity: 0;
@@ -132,5 +134,6 @@ export default function loadStyle() {
     styleTag.innerHTML = style;
     styleTag.id = 'expo-image-styles';
     document.head.appendChild(styleTag);
+  } else {
   }
 }
